requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.8.2
motor==2.5.1
sentry-sdk==1.4.3
python-jose==3.3.0
httpx==0.21.1

aioprometheus[starlette]==21.9.1

# AWS
boto3==1.20.26
botocore==1.23.26
requests-aws4auth==1.1.1
opensearch-py==1.0.0

# GCP
google-cloud-bigquery==2.31.0
